+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Skill roll chat cards now have a button to open the associated compendium entry
 - Add conditional modifiers to attack type items, due to the efforts of Ethaks (many thanks!)
 - Changed display of tags in basic item chat cards to look the same as those from action cards
+- Compendium browsers now remember their filters for the session
 
 ## 0.75.2